requirements.txt
fastapi
uvicorn[standard]
pydantic
onvif-zeep
zeep
requests
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
